requirements.txt
aiohttp==3.9.1
asyncpg==0.26.0
aioredis==1.3.1
pydantic==1.10.10
python-dotenv==0.14.0
Quart==0.17.0
quart_schema==0.11.1
shortuuid==1.0.11
uvicorn==0.20.0
Werkzeug==2.1.2